Renames base class file to not conflict with Jest.
